refac: separate svgoConfigBuilder tests from loader tests
