+++ test/loader.test.js
@@ -5,26 +5,7 @@ import compiler from './compiler.js';
 const fs = require('fs');
 import path from 'path';
 import vueSvgLoader from '../index.js';
-const sanitizeSvgoConfig = require('../svgoConfigBuilder.js');
 const semVerMajor = require('semver/functions/major');
-
-const config = {
-  plugins: [
-    {
-      name: 'preset-default',
-      params: {
-        overrides: {
-          inlineStyles: {
-            //force deletion of <style> tags in order for vue-loader to accept the template (no nested tags with side effects allowed)
-            onlyMatchedOnce: false
-          }
-        }
-      }
-    }
-  ]
-}
-
-const copy = obj => JSON.parse(JSON.stringify(obj));
 
 function getActualTemplate(stats){
   let source = stats.toJson({source: true}).modules[0].source; //get the compilation result
@@ -80,30 +61,3 @@ it('injects listeners for Vue 2', async () => {
   vueTemplate = vueSvgLoader.call(context, file);
   expect(vueTemplate).not.toContain('v-on="$listeners"');
 })
-
-it('should produce a valid svgo config', () => {
-  expect(sanitizeSvgoConfig({})).toEqual(config);
-});
-
-it('should not change a valid svgo config', () => {
-  expect(sanitizeSvgoConfig(config)).toEqual(config);
-});
-
-it('should override onlymatchedonce to false', () => {
-  const test = copy(config);
-  test.plugins[0].params.overrides.inlineStyles.onlyMatchedOnce = false;
-  expect(sanitizeSvgoConfig(test)).toEqual(config);
-});
-
-it('should keep unrelated information unchanged', () => {
-  const test = copy(config);
-  test.plugins[0].params.overrides.inlineStyles.onlyMatchedOnce = true;
-  test.plugins.push({name: "foobar", params: {foo: "bar", ...copy(config.plugins[0].params)}})
-  const expectation = copy(test);
-  expectation.plugins[0].params.overrides.inlineStyles.onlyMatchedOnce = false;
-  expect(sanitizeSvgoConfig(test)).toEqual(expectation);
-});
-
-it('should create a new config from a passed svgoConfig', () => {
-  expect(sanitizeSvgoConfig(config)).not.toBe(config);
-});
